requirements.txt
aiogram
python-dotenv
google-api-python-client
aiohttp